pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tabela-nutricional-web"
version = "0.1.0"
description = "ANVISA-compliant nutritional table calculator (RDC 429/2020, IN 75/2020)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Terracota" }]
keywords = ["anvisa", "nutritional-labeling", "rdc-429", "in-75"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "flask>=2.0",
    "gunicorn>=21.0",
    "flask-login>=0.6",
    "flask-sqlalchemy>=3.1",
    "flask-wtf>=1.2",
    "flask-limiter>=3.5",
    "flask-migrate>=4.0",
    "authlib>=1.3",
    "argon2-cffi>=23.1",
    "requests>=2.28",
    "email-validator>=2.0",
    "openpyxl>=3.0",
    "python-dotenv>=1.0",
    "python-json-logger>=2",
    "pytz>=2024.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = [".", "src"]
